pyproject.toml
[project]
name = "anchor-stack"
version = "0.1.0"
description = "AI-friendly engineering foundation with stable versions, unified logging, and pluggable capability packs"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Anchor Stack Team" }
]
keywords = [
    "mcp",
    "ai",
    "scaffolding",
    "template",
    "engineering",
    "debug"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"

dependencies = [
    # MCP SDK - Core dependency for MCP Server
    "mcp>=1.25.0,<2.0.0",

    # Data validation (already included by mcp, but explicit for our use)
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",

    # Template rendering
    "jinja2>=3.1.6,<4.0.0",

    # Configuration file parsing
    "pyyaml>=6.0.3,<7.0.0",

    # CLI support
    "click>=8.3.0,<9.0.0",

    # Typing extensions (already included by mcp)
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
anchor-stack = "anchor_stack.cli:main"

[project.urls]
Homepage = "https://github.com/anchor-stack/anchor-stack"
Documentation = "https://github.com/anchor-stack/anchor-stack#readme"
Repository = "https://github.com/anchor-stack/anchor-stack"
Issues = "https://github.com/anchor-stack/anchor-stack/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anchor_stack"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.isort]
known-first-party = ["anchor_stack"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src/anchor_stack"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
